+++ hw-ledger-transports/empty.ts
@@ -1,0 +1,8 @@
+// Copyright 2017-2022 @polkadot/hw-ledger authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import type { TransportDef } from './types.ts';
+
+export { packageInfo } from './packageInfo.ts';
+
+export const transports: TransportDef[] = [];