+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- @polkadot/hw-ledger-transports 9.5.2-13
 - @polkadot/hw-ledger 9.5.2-13
 - @polkadot/wasm-util 6.1.4
 - @polkadot/wasm-crypto-wasm 6.1.4